Add QIcon SVG specific tests to QtSVG module.
These tests used to reside in the qtbase module. As they rely on the
presence of the QtSVG module to execute they are not run as part of the
autotest suite. It makes more sense to place these autotest where they
can provide meaningful feedback to commits.

Task-number: QTBUG-24139
Task-number: QTBUG-22360

Change-Id: Ib7061598245673ac281e4b10cbe2243977264b6d
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>