Let svg plugins extend svg module
svg plugins are by default not linked together with
QtGui since its usage is rare. But when the pro file
contains Qt += svg, then its the opposite; The
application is likely to need the svg plugins.

Many of the examples shipped with the SDK fails building
on iOS (e.g accelbubble) because of this.

Change-Id: I83873d5405fde2b3d38e836057d55fcc2ceb9b6c
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>