Allow running qicon_svg test against local qtsvg build
When the build is local (i.e. not system-wide installed), libQt5Svg.so.5
needs to be on LD_LIBRARY_PATH, otherwise the plugin will not be able to
load it.

LD_LIBRARY_PATH is set by target_wrapper.sh, generated by qmake, which
will add local lib directory there only if QT contains svg.

This is a workaround for the dependency resolution of plugins and
modules not being properly integrated.

Change-Id: I68c3bc850367af62e7ccaa766fa83453fa0d8621
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>