QSvgIconEngine::addFile(): Use QMimeDatabase to determine file type
Introduce an enumeration to describe the file type and use
QMimeDatbase to determine the type. This will detect SVG files
also in the case resource aliases without suffixes are used.
As a drive by, do not pass compressed SVG files
to QPixmap when QT_NO_COMPRESS is defined, as this likely
results in garbage.

Task-number: QTBUG-55388
Change-Id: Ib8bafca60e9238412a5c7edf560d487e512b8fd0
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>