Eradicate Q_FOREACH loops and mark the module as Q_FOREACH-free
In BearerCloud::timerEvent(), bite the bullet and use a std::vector
instead of a QList to avoid repeated qAsConst() applications.

In BearerCloud::configurationRemoved() and configurationChanged(),
instead of iterating over QMultiMap::uniqueKeys() and QMap::remove(),
iterate directly over the QMultiMap and remove the desired elements.
Cache a QString.

In BearerCloud::updateConfigurations(), instead of using while !isEmpty()
takeFirst() on a local copy of a QList, simply iterate over the QList
(now made const) using C++11 ranged for.

In tst_QSvgGenerator, replace a QList of statically-known size with
a C array.

Change-Id: Ic0dd1c67d0819fe6167f2bce248f9b910be65803
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>