Make svg icons work as pixmap icons
This patch makes the svg icon code similar to the code for pixmap icons.
We only style the code if we fetch a different mode (+it is not normal)
and we try to find icons in same order as the pixmap icons.

[ChangeLog][][QIcon/QSvgIconEngine] Made SVG icons behave like pixmap icons:
An explicitly-set disabled icon is no longer additionally grayed out.

Task-number: QTBUG-67452
Change-Id: I94b4146bae2e9d924a96b602bf23909bd7f0e934
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Morten Kristensen <msk@nullpointer.dk>
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>
Reviewed-by: Andy Shaw <andy.shaw@qt.io>