Convert uses of QTime as a timer to QElapsedTimer
One was an easy replacement, the other changes its reference time,
which isn't supported by QElapsedTimer. However, it was easy enough
to reimplement in terms of QDateTime::currentMSecsSinceEpoch(); and,
as it was previously using QTime as timer, this is presumably good
enough (although a proper monotonic timer would be better).

Change-Id: I90f8e08635e7d2bc364755efd93fce008b6a62c0
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>