Fix rendering of SVG with translating viewBox
This is a modification of14fa4591eb34a35cf3d485fd901e3f1e2caa7770.

That change correctly implemented support for scaling and centering
viewBox. However, the viewBox scaling and x-min/y-min translation was
applied in the wrong order, so SVGs with a non-zero x-min/y-min
viewBox would be rendered with an offset.

The autotest reflected this error, and is also fixed here. (In user
space, the black rectangle's upper left corner (0,0) is in the center
of the viewBox and so should end up in the center of the produced
image, independently of the scaling. All other tested SVG renderers
place it there.)

This change fixes all the reported regression cases from the above
commit, while not re-breaking the original cases fixed by that commit.

Fixes: QTBUG-79333
Fixes: QTBUG-78200
Task-number: QTBUG-70256
Change-Id: I0d3394e3caf6ec7edf16a10992c2fbfdac7398e5
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>