CMake: Handle zlib dependency correctly
Unfortunately whenever a module or plugin uses zlib, the .pro file
contains a copy-pasted conditional scope when to use system zlib and
when to use the Qt bundled zlib.
The pro2cmake script handles the linking part, but not the
find_package part.
We need to explicitly search for the correct package depending
on whether system zlib feature is on or off, and also make sure
not to search for the system one in case it already exists (can
happen in static builds).

Change-Id: I9c0d678815e4dc0f8e239c3b4667a375aeb757c1
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>