Fix wrong calculations in animated transforms
In the implementation of the AnimateTransform node, the
percentOfAnimation variable is clearly intended to hold the current
instant's local fraction between animation step n and n+1. Instead it
held the global fraction between step 0 and final step.

Fix the calculation, and also rename the variables to make it easier
to follow.

Change-Id: Iefcb5cdf76d838de34e5a33ee4fd78721fa698fd
Done-with: Jean-Milost Reymond
Task-number: QTBUG-82779
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>