Update commercial license headers
Updated header.COMM to the files in tqtc-qtsvg. Examples, tests or
documentation files are not updated.

The commercial license header may contain
some additional lines so that its line count
equals with the earlier license header. Reason
for this is that some autotests use hard coded line
numbers and a change in the line count causes
failures in tests.

Task-number: QTQAINFRA-4162
Change-Id: Iedfe87d55dc1967f8340101279c647f1cf3534ca
Reviewed-by: Antti Kokko <antti.kokko@qt.io>