Remove .qmake.conf from this repository
This file was necessary for the qmake build and the CI.  The qmake
build is history, and the CI does now read .cmake.conf (see
QTQAINFRA-4392).

In addition to being superfluous, the existence of .qmake.conf
triggers QTBUG-76140 when building Qt examples with qmake.  Removing
the file alleviates the symptoms of this bug.

Fixes: QTBUG-92271
Change-Id: Ib55f459676666817185305a69040620c0d6d1bf6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>