Improve <use> node resolution process
Since QSvgUse nodes are the only ones that ever get stored in the
m_resolveNodes member, we can avoid extra cast and checks around it.
Also a more descriptive name wouldn't hurt.

Currently, the list is merely cleaned at the end of a parsing process,
but ideally it should not be stored at all, hence I left the comment
for possible future optimizations, if anyone's willing to do it.

Pick-to: 6.2
Change-Id: I86e83a8661679503e532c230e70f6d52092ad58e
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>