Fix missing null-check in parseBaseGradient()
Of the four functions returned by findStyleFactoryMethod(), two do in
fact take care to cope if the first parameter is null; but the other
two forward it to parseBaseGradient(), which did not, provoking a
warning from CodeChecker.

The single caller of parseBaseGradient() is in an if/else cascade in
which some branches do assert m_nodes.isEmpty() before passing
m_nodes.top() to their respective functions, but the
parseBaseGradient() branch did not. I infer that its returns should
check for null and duly add the missing null check, rather than
asserting against it.

Change-Id: I6c630f04522a0ae05021fe95d848be64abe6c20c
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>