Update commercial license headers
Updated header.COMM to the files in tqtc-qtsvg.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4942
Change-Id: Id957855460f5e2a40a4ca091d32a3daeea69b095
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>