Let CSS attributes have higher priority than the plain XML attributes
Although seemingly not well specified, the common behavior of web
browsers and other SVG renderers is to give priority to attribute
values specified in the CSS style attribute, in case of conflict. So
make qtsvg do the same, by moving the CSS parsing block to after the
XML attribute parsing block.

Fixes: QTBUG-100068
Pick-to: 6.4
Change-Id: I769303d69509511103d96e8d26e2cfeac1f19494
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>