QSvgPaintEngine: code tidies
When the paint engine state changes and updateState gets called,
the engine unconditionally streams out the entire gstate.
To do so, the dirty flags are OR'ed with a all-1 mask, and then
its flags individually tested. But those tests are now always true,
because *we* just set every flag. Just remove the flag handling.

Change-Id: Ibd9bbaef91f189da5366bc24498926c3d3abed20
Reviewed-by: Albert Astals Cid <aacid@kde.org>