Port from qAsConst() to std::as_const()
We've been requiring C++17 since Qt 6.0, and our qAsConst use finally
starts to bother us (QTBUG-99313), so time to port away from it
now.

Since qAsConst has exactly the same semantics as std::as_const (down
to rvalue treatment, constexpr'ness and noexcept'ness), there's really
nothing more to it than a global search-and-replace.

This is a 6.4 re-run of the script we ran in dev, in order to avoid
conflicts between the branches when cherry-picking.

Change-Id: I764d628eae9c657787079bf88b0b8a30d97e5fce
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>