Don't rasterize gigantic shapes
We set a limit for the bounding rect of shapes to avoid stalling
applications for several minutes while rendering unreasonably large
shapes. The limit can be disabled using an environment variable.

[ChangeLog][Performance] Unreasonably large shapes are now ignored
in SVG files to avoid stalling the application. The check can be
disabled by setting QT_SVG_DISABLE_SIZE_LIMIT=1 in the environment.

Pick-to: 6.2 6.5
Fixes: QTBUG-111850
Change-Id: Id8154049c96a565aad237ee007da0ee879446448
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>