Add mask element and mask attribute to QSvg
The QSvgHandler can now parse the mask element and the mask attribute of
all other elements. If a mask is set on an element, the rendering is
redirected into a buffer and the mask is applied to this buffer before
it is drawn onto the canvas.

Task-number: QTBUG-115223
Task-number: QTBUG-115543
Change-Id: I14b2966f7b9ee2717c8f29f9090f986c2a968c3a
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>