Add currentColor support to gradient <stop/> node
This commit combines a refactoring and a fix/feature.

The removed code was a partial duplicate of the parseCSStoXMLAttrs()
function, which is called from cssStyleLookup() for in most other places.
Other places also have these #ifndef guards and an additional
parseStyle() call after that. In turn, parseStyle() calls parseColor()
which is exactly what was missing to push color attribute on the stack,
so that constructColor()/resolveColor() could query it from the handler
next.

All in all, this allows gradient stop to use classes with only color
CSS property specified, thus removing the need for frameworks to bloat
injected CSS with duplicating and rarely used stop-color properties.

I used Plasma Framework's "widgets/glowbar" SVG for testing, and with
some basic modifications it now properly reacts to Plasma accent color.

Change-Id: Ic9b96836280a8ef7a0b34f0401cd64dac99370f3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Marco Martin <mart@kde.org>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>