Fix crash when reading text with line breaks
Introduced by d130c74f4f54cddcd5baa7d8dc4c5b68772345c2, trying to read
an svg with a line break in the text would crash, because nullptr is
used as a token for linebreaks in Qt SVG and we were accessing it without
checking.

[ChangeLog] Fixed a regression where the application would
crash on certain SVG files.

Pick-to: 6.2 6.5 6.6
Change-Id: Ia261372cd957a2ebcd49033a57d43ebd201c83bb
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>