Refactor QSvgNode draw architecture
In preparation for the mask and filter tag, the draw infrastructure
is changed from a C-style macro to virtual member functions.
Commonalities between all nodes are encapsulated in the draw()
function. The actual draw command is encapsulated in the
drawCommand() function that is called by the draw function along with
some common code for setting styles, masking, filtering and
benchmarking.

Change-Id: I19aa79be9500eff6faa6b6f12fd13d6c0fe2d0ad
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>