QSvgWidget: only start the animation timer when visible
Thanks to Gammaray, I noticed that a hidden (and rarely shown) QSvgWidget
whose purpose is to show a "searching..." animation was triggering a 33ms
QTimer for ever in the application. The slot simply calls update()
which doesn't do anything on a hidden widget, but waking up every 33ms
is still wasteful -- and noisy in gammaray's Events module.

[ChangeLog] Add animationEnabled property to QSvgRenderer,
and use it from QSvgWidget to stop animations when hidden.

Change-Id: I4ec63254b49a513688f8274cee49a1251b89a6cd
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>