Remove nodeStringList in favor of QSvgNode::typeName()
When adding node types (e.g. mask) the QSvgStyleSelector_nodeString
was not updated. With this patch we use QSvgNode::typeName()
instead. The return values were adjusted to ensure that we are
backwards compatible.

This list is required to make CSS work properly. The missing strings
triggered a crash when testing the resvg test suit. This patch fixes
the crash.

Pick-to: 6.7
Change-Id: Ia715664c87d531b125627554129b8e4d930ced44
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>