Allow opting out of resource limits on parsing and rendering
Some checks against corrupt/crafted svgs can inhibit loading unusual
but legal svg files. Add a mechanism for disabling such checks for
loading trusted svgs, and implement it for the maximum limit on the
number of points in an svg path element.

Initially only settable by environment variable, so it can be
cherry-picked to earlier branches. To be followed up with a new Option
flag that can be specified on the QSvgRenderer level in new Qt
versions.

Fixes: QTBUG-120653
Pick-to: 6.7 6.6 6.5
Change-Id: I1c1a3f31373d6f6e6ba288bb7c907ccb6f0b3611
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>