Fix a bug where description and title were rendered
Parts of SvgHandler wrongfully claim that description and title are
handled. However, this is not the case and the current element is not
invalidated. This in turn leads to the acceptance of the text in the
respective element which is added to the parent and in case of text
rendered.

This patch removes the claim that these elements are handled which is
enough to fix the bug.

Fixes: QTBUG-120233
Pick-to: 6.7
Change-Id: I60eab19c96a556647c3e96526f7d1135cbdac91e
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>