Replace FeatureSet with Options, a more flexible flags parameter
The featureSet enum setting was added for 6.7, but discussions around
API review have commented on it as somewhat inflexible, and we have
also already identified expected needs for other, similar settings not
covered by the featureset concept. For example, more fine-grained
control on enabling/disabling of certain features, or a "trusted
source" setting that disables certain safety checks.

To avoid more API churn, both in the public and all the constructors
of the private classes where this setting is used, this commit
replaces FeatureSet with Options, a QFlags type where new bitflags can
easily be added in the underlying Option enum as needed in the future.

Change-Id: I07b5cbde2de45f7c8dabeace2b491a11126207be
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 1868b3f6ed10a051dcd0138f580a6b99738361ee)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>