Enable <use> elements in <mask> and propagate state to mask children
The use node could not be created within a mask due to a missing entry
in a switch statement. This is corrected with this patch.

The mask element was implemented such that states do not propagate to
it's children. This is not correct and states are propagated correctly
with this patch.

Pick-to: 6.7 6.7.0
Fixes: QTBUG-123010
Change-Id: I4aab070c73b2e8a9bd0c642c70ea5b28eaef3416
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>