Enable <use> elements in <mask> and propagate state to mask children
The use node could not be created within a mask due to a missing entry
in a switch statement. This is corrected with this patch.

The mask element was implemented such that states do not propagate to
it's children. This is not correct and states are propagated correctly
with this patch.

Fixes: QTBUG-123010
Change-Id: I4aab070c73b2e8a9bd0c642c70ea5b28eaef3416
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 247453e7d5c84837b80444c26b1afd6165f63477)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 5b7240d963f22988bd08a78941de8e547ef35946)