Avoid runtime warnings about negative font sizes
The calculations assume that fonts have a point size, and not a pixel
size set. But if an svg was rendered to a paint device (eg. a widget)
that had a pixel sized font, runtime warnings would be produced, since
the point size in that case is reported as -1.
Fix by ensuring the painter is initially reset to a default, point
sized font.

Fixes: QTBUG-122752
Pick-to: 6.6 6.5 5.15
Change-Id: I5cfbb95ac3f5ac1299ae8a111bf8278836bf397b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit 863929f44639203c6678ca0f1a16e69e2d941cf0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>