Fix filter offset with transformations
There was a missconception in the calculation of localFilterBoundingBox
that lead to a rendering error of filter elements if there was a
transformation set.

The bounding box is always positioned relative to the node that the
filter is applied to. This patch corrects the calculation of the
bounding box.

Fixes: QTBUG-123138
Fixes: QTBUG-127018
Change-Id: I9ed5ed3a2bb55bc61002cd0c506c1f2310962156
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit c58e6c9f276c1c524a3d53a41ee953e3b763b2eb)
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>