Filters: check image allocation to avoid memory exhaustion and crash
The new svg features introduced in 6.7 generally use the checked
QImage allocation from ImageIO to create the intermediate images. The
exception was in applyFilter(), where a QImage was created by the copy
function.

This commit does not fix the underlying problem in the filter size
calculation. It just ensures that the code fails cleanly, whatever the
reason for the unexpected size.

Pick-to: 6.7
Task-number: QTBUG-124287
Change-Id: If1a20b11842cc05645fe16aa9f4158190f28f5b2
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>