Support loading image file whose name begin with "data"
We would assume all filenames beginning with "data" indicated
embedded base64, and even if this failed, we would not try
loading the file as normal. This made e.g. loading a file called
"data.png" impossible.

The easy fix is just to check if we successfully loaded the file
as data and load it as a file instead if not.

Change-Id: I19e532ae7a8fd456ba9717c36820a188a2cec470
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>