Skip rendering filters with unsupported primitives
In Qt 6.7, <filter> and some of the filter primitives were supported.
However, this caused a regression in behavior because Qt Svg used to
skip filters altogether, but starting from Qt 6.7, the filter and the
filter primitives are parsed and the unsupported filter primitives are
skipped which leads to unexpected output.

It is a good idea to start supporting the rest of the filter primitives
in the future, but for now the renderer will skip applying the filter if
it has an unsupported filter primitive.

Fixes: QTBUG-123994
Pick-to: 6.7
Change-Id: I14b91c919ace8a886ee285ff1ed466110d13322a
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>