Fix transformed Gaussian blur filter
The Gaussian blur filter is applied after the item is transformed to
ensure that the effect is resolved well and no single pixels can be
seen.

However, this leads to errors when the item is rotated or skewed,
since the gaussian blur effect has to be applied in the local
coordinate system (see stdDeviation description in standard).

This patch decomposes the transformation matrix into a scaling and a
remaining component. The blur effect is applied in the local but
scaled coordinate system to ensure a high resolution of the effect.
The result is then transformed with the remaining component of the
transformation matrix to get the final fitered item.

Change-Id: I48f1837e5f34fef3bccf7d537a458bed857067b0
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit ba44960d8f2458b829a4d4ffbdb2dd1748d77857)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>