Revert style recursive after it has been applied recursive
The mask and pattern node have a drawing hierarchy that does not match
the document hierarchy, as they can be called from any point in the
document.

Therefore we have to reconstruct the painter state recursivly (see
commit 247453e7d5c84837b80444c26b1afd6165f63477). However, we also have
to revert the style recursivly which is done in this patch.

Pick-to: 6.7
Change-Id: I939595182674f8aed6b265bc633476e78807727c
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
(cherry picked from commit cffa99bdffbf31820c7ad29e9903966be2194759)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>