Do not try to optimize filter size
In the original filter patch I tried to be smart and reduce the buffer
size by calculating the required rectangle and limiting the buffer to
it.

This does not work with a pipeline of multiple filters, as it cannot be
predicted what the following filters need. Therefore this change
removes the respective code.

Pick-to: 6.7
Change-Id: If765f9b5e4fc8db186e2c2dba7c16e5517bec6e8
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit bc9b5de9f9a0f1da2ad82ad861ccfb1256a8b41c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>