Improve Gaussian blur filter performance
The Gaussian blur filter is based on a repeated application of the box
blur filter. This filter requires a large amount of average pixel values
in a subarea. The old implementation used a naive implementation that
scales with O(n*r²) where n is the pixel number and r the blur radius.
This is a considerable problem and can take the majority of the render
duration of an svg image. The blur radius r often grows with the zoom
factor and can thus become very large.

The algorithm can be optimized to O(n) with some smart buffering of pixel
value sums. Then the algorithm is independent of the blur radius and
zooming has a much lower impact on the rendering duration.

This patch applies the optimization and the Gaussian blur filter now
scales with O(n). With help from Jan-Arve Sæther.

Some simple tests with the old and new code showed the following render
duration (only the application of the filter):

Test           |   old version    |   new version
small picture  |       0.5 ms     |      0.2 ms
large picture  |       330 ms     |      135 ms
large radius   |     20700 ms     |      160 ms

Pick-to: 6.7 6.8
Change-Id: I886965efe7121e7afea9aaca657aeceb8ece7c63
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>