Expose some animation properties
This exposes a few properties for enabling use cases such
as VectorImage to interpret the animations.

This also fixes a bug where animateColor would be reverted on
a painter without first being applied, which meant reverting it
to a default pen/brush because the old one had never been saved.
We fix this the same was QSvgAnimateTransform handles it: by
always saving the old values even if the style is not applied.

Task-number: QTBUG-127590
Change-Id: If7460bdb4ff492ad72765253ec773ad07ef642a0
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>