Fix a crash in transform animation
Popping an empty stack fires an assertion which is a regression
introduced by changes to QSvgTransformStyle to hold a stack of
transformations to apply and revert it recursively when used with mask
and pattern. This is a hotfix for the issue which is not done on the dev branch, because the refactoring work for animations will fix this issue.

Fixes: QTBUG-128485
Change-Id: I2fb36a168ceb2acc85b5aab59ca081ee5b934553
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>