Extend benchmark for rendering
With the exended features it is important to have an eye on rendering
performance. Both rendering on a QPicture and QImage are tested to
isolate the processing in QSvgRenderer and QPainter.

Tested are standard path drawing with tiger.svg as well as elements that
operate on pixel buffers, that are filters, mask and the correct
implementation of group opacity.

Manually cherry-picked from 90114ae5da84714b5c5ec88151a3bf133363c43f

Task-number: QTBUG-135075
Pick-to: 6.8
Change-Id: I2a877c72f6000725f1b4372a04184c26e5c74243
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Robert Löhning <robert.loehning@qt.io>