Fix issues with stroke animations
There are two issues solved by this commit. First, the wrongly
generated number of key frames in the case an animateColor is used
which led to incorrect animations when the "values" attribute is
defined. Second, stroke animations did not run if there was no
initial stroke value defined. This happened because the brush
style was set to NoBrush by QSvgStrokeStyle::apply

Fixes: QTBUG-135596
Pick-to: 6.9
Change-Id: I6ee9c686b1d60f6cf754cac749b3dce5185e2bfa
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>