Respect transform order in CSS animation
Any order of transformation functions can be declared in "transform"
property in a CSS style, which means they have to be applied one
after another from right to left. Moreover, the same function can be
declared more than once in that list and they have to be taken into
account. QSvgAnimator did not respect this order and always collected
all the values and applied transformations in this order :
Skew->Scale->Rotate->Translate

Change-Id: Ibf598431d451240d307dc5adccb5786c22216ca9
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>