Parse CSS animation properties separately
CSS animation properties were stored in QSvgAttributes class, but
those properties are not part of SVG standard and furthermore,
they have to be handled differently because they have to be parsed
in the order they appear in a CSS style sheet to allow shorthand
and longhand form to override each others when both are defined at
the same time. This commit prepares Qt Svg to be able to handle
shorthand form of CSS animation properties.

Change-Id: Ib294c9722f599e96340f595f940cdf7fd041c070
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>