Parse multiple CSS functions in a declaration
When the handler parses a list of CSS declarations, it assumed that
there would be only a single function in the list of values for a
given declaration. This is incorrect as some properties like transform
or animation-timing-function can accept multiple functions at the same
time.

Change-Id: Ia537a52f5ead84c023ac9e62bff681e3c2a0f892
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 6065c5fede827a8bd595f18c017b2a549fbb28c0)