CMake: Make qtsvg benchmark test a standalone project
Add the boilerplate standalone test prelude to the test, so that it
can be opened with an IDE without the qt-cmake-standalone-test script,
but directly with qt-cmake or cmake. Boilerplate was copied from the
auto tests which already had it.

I also had to remove the requirement of Qt::Widgets which seemed unused
anyway.

Pick-to: 6.9 6.8
Change-Id: I19d49c8d2373eb325b9eb7936997033471584d18
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>