Fix rendering opacity in text elements
For rendering images, qtsvg uses QPainter which only knows one global
opacity. To have different opacities for fill and stroke, a text needs
to be rendered twice: Once only drawing the outline and once only
drawing the foreground fill. This has already been the case for other
elements but not for text yet.

[ChangeLog] Fixed rendering fill-opacity and stroke-
opacity in text elements

Fixes: QTBUG-126268
Pick-to: 6.10
Change-Id: Iba7fcc059480bbdc8fc5a4f652eb653f45e66ffb
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>