Move functions to qsvgutils
Some functions that are used as helpers to parse an svg path will
be used by the CSS Handler to parse the path defined for animations.
Those functions are moved to the QSvgUtils namespace so it can be
reused. Moreover, parsePathDataFast function now returns an
std::optional instead of relying on the ok bool.

Change-Id: Ib5da7675e8539958c602fb3c2188677e26801f65
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>